pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crtr"
version = "0.1.0"
description = "Convert Instagram cooking reels into structured recipes using AI"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Thomas Larsen", email = "Larsen.thomas.dk@gmail.com"}
]
keywords = ["instagram", "recipe", "ai", "video", "transcription", "cooking"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Multimedia :: Video",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "instaloader>=4.10.0",
    "requests>=2.31.0",
    "moviepy>=1.0.3",
    "faster-whisper>=1.0.0",
    "torch>=2.0.0",
    "google-genai>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/crtr"
Repository = "https://github.com/yourusername/crtr"
"Bug Tracker" = "https://github.com/yourusername/crtr/issues"

[project.scripts]
crtr = "crtr.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
crtr = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
